Update windows-10-feature-updates.md
I'm not certain this is absolutely accurate.  The "Feature updates" policy in Intune doesn't allow anything prior to 1909 today (unless I'm looking at something wrong).  I'm assuming that we're going to restrict "pinning" to only supported versions of the OS.  Might make sense to take the version callout out of this doc entirely and explain it as "...that you want devices to remain at.  This only includes currently supported versions of Windows 10."  Maybe even make it a purple note so that it's clearer.